requirements.txt

streamlit>=1.32
pandas>=2.2
numpy>=1.26
google-genai>=1.55
rapidfuzz>=3.0.0
